abstract common utilities into an api_utils.py file
